For OS_WIN32, 'pid_t' is now typedeffed as 'int'.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3707 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_signal.h
@@ -34,7 +34,7 @@ extern "C" {
 
 #if defined(OS_WIN32)
 
-typedef unsigned long pid_t;
+typedef int pid_t;
 typedef unsigned int uid_t;
 typedef unsigned long sigset_t;
